Fixing id types from int to long

+++ LivestockFarmCommon/src/main/java/com/co/livestockFarm/dto/FoodDTO.java
@@ -3,6 +3,6 @@ import lombok.Data;
 
 @Data
 public class FoodDTO {
-	public int foodId;
+	public long foodId;
 	public String name;
 }
